move statistics clock from to GeodeRedisServerRule to test instance
