+++ geode-redis/src/commonTest/java/org/apache/geode/redis/GeodeRedisServerRule.java
@@ -24,8 +24,6 @@ import static org.apache.geode.distributed.ConfigurationProperties.REDIS_PASSWOR
 import org.apache.geode.cache.CacheFactory;
 import org.apache.geode.cache.GemFireCache;
 import org.apache.geode.internal.cache.InternalCache;
-import org.apache.geode.internal.statistics.EnabledStatisticsClock;
-import org.apache.geode.internal.statistics.StatisticsClock;
 import org.apache.geode.redis.internal.GeodeRedisServer;
 import org.apache.geode.test.junit.rules.serializable.SerializableExternalResource;
 
@@ -33,8 +31,6 @@ public class GeodeRedisServerRule extends SerializableExternalResource {
   private GemFireCache cache;
   private GeodeRedisServer server;
   private CacheFactory cacheFactory;
-  private final long START_TIME;
-  private StatisticsClock statisticsClock;
 
 
   public GeodeRedisServerRule() {
@@ -42,8 +38,6 @@ public class GeodeRedisServerRule extends SerializableExternalResource {
     cacheFactory.set(LOG_LEVEL, "warn");
     cacheFactory.set(MCAST_PORT, "0");
     cacheFactory.set(LOCATORS, "");
-    statisticsClock = new EnabledStatisticsClock();
-    START_TIME = statisticsClock.getTime();
   }
 
   @Override
@@ -53,13 +47,7 @@ public class GeodeRedisServerRule extends SerializableExternalResource {
     server.setAllowUnsupportedCommands(true);
   }
 
-  public long getStartTime() {
-    return START_TIME;
-  }
 
-  public long getCurrentTime() {
-    return this.statisticsClock.getTime();
-  }
 
   public GeodeRedisServerRule withProperty(String property, String value) {
     cacheFactory.set(property, value);
